move chaos evaluator folder
