+++ src/strands_evals/evaluators/chaos/failure_communication_evaluator.py
@@ -5,14 +5,14 @@ from pydantic import BaseModel, Field
 from strands import Agent
 from strands.models.model import Model
 
-from ...evaluators.evaluator import Evaluator
 from ...types.evaluation import EvaluationData, EvaluationOutput, InputT, OutputT
 from ...types.trace import EvaluationLevel
-from .prompt_templates.recovery_strategy import get_template
+from ..evaluator import Evaluator
+from .prompt_templates.failure_communication import get_template
 
 
-class RecoveryStrategyScore(str, Enum):
-    """Categorical recovery strategy ratings."""
+class FailureCommunicationScore(str, Enum):
+    """Categorical failure communication ratings."""
 
     FAILURE = "Failure"
     POOR = "Poor"
@@ -21,24 +21,24 @@ class RecoveryStrategyScore(str, Enum):
     EXCELLENT = "Excellent"
 
 
-class RecoveryStrategyRating(BaseModel):
-    """Structured output for recovery strategy evaluation."""
+class FailureCommunicationRating(BaseModel):
+    """Structured output for failure communication evaluation."""
 
     reasoning: str = Field(description="Step by step reasoning to derive the final score")
-    score: RecoveryStrategyScore = Field(description="Categorical recovery strategy rating")
+    score: FailureCommunicationScore = Field(description="Categorical failure communication rating")
 
 
-class RecoveryStrategyEvaluator(Evaluator[InputT, OutputT]):
-    """Evaluates appropriateness of agent's recovery strategy when handling failures."""
+class FailureCommunicationEvaluator(Evaluator[InputT, OutputT]):
+    """Evaluates quality of agent's failure communication and user experience."""
 
     evaluation_level = EvaluationLevel.TRACE_LEVEL
 
     _score_mapping = {
-        RecoveryStrategyScore.FAILURE: 0.0,
-        RecoveryStrategyScore.POOR: 0.25,
-        RecoveryStrategyScore.ACCEPTABLE: 0.5,
-        RecoveryStrategyScore.GOOD: 0.75,
-        RecoveryStrategyScore.EXCELLENT: 1.0,
+        FailureCommunicationScore.FAILURE: 0.0,
+        FailureCommunicationScore.POOR: 0.25,
+        FailureCommunicationScore.ACCEPTABLE: 0.5,
+        FailureCommunicationScore.GOOD: 0.75,
+        FailureCommunicationScore.EXCELLENT: 1.0,
     }
 
     def __init__(
@@ -53,7 +53,7 @@ class RecoveryStrategyEvaluator(Evaluator[InputT, OutputT]):
         self.system_prompt = system_prompt if system_prompt is not None else default_prompt
         self.model = model
 
-    def _build_output(self, rating: RecoveryStrategyRating) -> list[EvaluationOutput]:
+    def _build_output(self, rating: FailureCommunicationRating) -> list[EvaluationOutput]:
         normalized_score = self._score_mapping[rating.score]
         return [
             EvaluationOutput(
@@ -68,14 +68,14 @@ class RecoveryStrategyEvaluator(Evaluator[InputT, OutputT]):
         parsed_input = self._get_last_turn(evaluation_case)
         prompt = self._format_trace_level_prompt(parsed_input)
         evaluator_agent = Agent(model=self.model, system_prompt=self.system_prompt, callback_handler=None)
-        result = evaluator_agent(prompt, structured_output_model=RecoveryStrategyRating)
-        rating = cast(RecoveryStrategyRating, result.structured_output)
+        result = evaluator_agent(prompt, structured_output_model=FailureCommunicationRating)
+        rating = cast(FailureCommunicationRating, result.structured_output)
         return self._build_output(rating)
 
     async def evaluate_async(self, evaluation_case: EvaluationData[InputT, OutputT]) -> list[EvaluationOutput]:
         parsed_input = self._get_last_turn(evaluation_case)
         prompt = self._format_trace_level_prompt(parsed_input)
         evaluator_agent = Agent(model=self.model, system_prompt=self.system_prompt, callback_handler=None)
-        result = await evaluator_agent.invoke_async(prompt, structured_output_model=RecoveryStrategyRating)
-        rating = cast(RecoveryStrategyRating, result.structured_output)
+        result = await evaluator_agent.invoke_async(prompt, structured_output_model=FailureCommunicationRating)
+        rating = cast(FailureCommunicationRating, result.structured_output)
         return self._build_output(rating)
