+++ src/strands_evals/evaluators/chaos/prompt_templates/partial_completion/__init__.py
