+++ src/strands_evals/evaluators/chaos/prompt_templates/partial_completion/partial_completion_v0.py
