+++ tests/strands_evals/chaos/test_recovery_strategy_evaluator.py
@@ -5,8 +5,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from strands_evals.chaos.evaluators import RecoveryStrategyEvaluator
-from strands_evals.chaos.evaluators.recovery_strategy_evaluator import (
+from strands_evals.evaluators.chaos import RecoveryStrategyEvaluator
+from strands_evals.evaluators.chaos.recovery_strategy_evaluator import (
     RecoveryStrategyRating,
     RecoveryStrategyScore,
 )
@@ -85,7 +85,7 @@ def test_init_with_custom_values():
     assert evaluator.system_prompt == "Custom"
 
 
-@patch("strands_evals.chaos.evaluators.recovery_strategy_evaluator.Agent")
+@patch("strands_evals.evaluators.chaos.recovery_strategy_evaluator.Agent")
 def test_evaluate(mock_agent_class, evaluation_data):
     mock_agent = Mock()
     mock_result = Mock()
@@ -116,7 +116,7 @@ def test_evaluate(mock_agent_class, evaluation_data):
         (RecoveryStrategyScore.FAILURE, 0.0, False),
     ],
 )
-@patch("strands_evals.chaos.evaluators.recovery_strategy_evaluator.Agent")
+@patch("strands_evals.evaluators.chaos.recovery_strategy_evaluator.Agent")
 def test_score_mapping(mock_agent_class, evaluation_data, score, expected_value, expected_pass):
     mock_agent = Mock()
     mock_result = Mock()
@@ -134,7 +134,7 @@ def test_score_mapping(mock_agent_class, evaluation_data, score, expected_value,
 
 
 @pytest.mark.asyncio
-@patch("strands_evals.chaos.evaluators.recovery_strategy_evaluator.Agent")
+@patch("strands_evals.evaluators.chaos.recovery_strategy_evaluator.Agent")
 async def test_evaluate_async(mock_agent_class, evaluation_data):
     mock_agent = Mock()
     mock_result = Mock()
